# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EpisodeStat": {
        "additionalProperties": false,
        "description": "Event counts for one contestant in one episode. One record per pair.",
        "properties": {
          "contestant_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
            "title": "Contestant Id",
            "type": "string"
          },
          "episode": {
            "exclusiveMinimum": 0.0,
            "title": "Episode",
            "type": "integer"
          },
          "events": {
            "additionalProperties": {
              "minimum": 0.0,
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/EventType"
            },
            "title": "Events",
            "type": "object"
          },
          "season_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
            "title": "Season Id",
            "type": "string"
          }
        },
        "required": [
          "season_id",
          "episode",
          "contestant_id"
        ],
        "title": "EpisodeStat",
        "type": "object"
      },
      "EpisodeStatIn": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "additionalProperties": {
              "minimum": 0.0,
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/EventType"
            },
            "title": "Events",
            "type": "object"
          }
        },
        "title": "EpisodeStatIn",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "EventType": {
        "description": "Things that happen to a castaway in an episode and earn (or cost) points.",
        "enum": [
          "survived_episode",
          "individual_immunity",
          "team_immunity",
          "individual_reward",
          "team_reward",
          "idol_found",
          "idol_played_successfully",
          "advantage_found",
          "correct_vote",
          "vote_received",
          "fire_making_win",
          "voted_out",
          "quit_or_medevac",
          "final_three",
          "jury_vote_received",
          "sole_survivor"
        ],
        "title": "EventType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Season": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "number": {
            "exclusiveMinimum": 0.0,
            "title": "Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "number"
        ],
        "title": "Season",
        "type": "object"
      },
      "SeasonIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "number": {
            "exclusiveMinimum": 0.0,
            "title": "Number",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "number"
        ],
        "title": "SeasonIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "fantasy-survivor-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/seasons/{season_id}": {
      "get": {
        "operationId": "get_season_seasons__season_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Season Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Season"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed for this user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "League rule violated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Authentication not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Season",
        "tags": [
          "seasons"
        ]
      },
      "put": {
        "operationId": "upsert_season_seasons__season_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Season Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeasonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Season"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed for this user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "League rule violated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Authentication not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Season",
        "tags": [
          "seasons"
        ]
      }
    },
...
    "/seasons/{season_id}/episodes/{episode}/stats": {
      "get": {
        "operationId": "list_episode_stats_seasons__season_id__episodes__episode__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Season Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "episode",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Episode",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EpisodeStat"
                  },
                  "title": "Response List Episode Stats Seasons  Season Id  Episodes  Episode  Stats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed for this user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "League rule violated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Authentication not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Episode Stats",
        "tags": [
          "stats"
        ]
      }
    },
    "/seasons/{season_id}/episodes/{episode}/stats/{contestant_id}": {
      "put": {
        "operationId": "record_stat_seasons__season_id__episodes__episode__stats__contestant_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Season Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "episode",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Episode",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "contestant_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Contestant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeStatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeStat"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed for this user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "League rule violated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Authentication not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Stat",
        "tags": [
          "stats"
        ]
      }
    },
...
    "/seasons/{season_id}/stats": {
      "get": {
        "operationId": "list_season_stats_seasons__season_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "season_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Season Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EpisodeStat"
                  },
                  "title": "Response List Season Stats Seasons  Season Id  Stats Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not allowed for this user"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "League rule violated"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Authentication not configured"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Season Stats",
        "tags": [
          "stats"
        ]
      }
    }
  }
}